wifi: mt76: move wed common utilities in wed.c
Introduce wed.c in order to collect mt76 wed common codebase used by
mt7915 and mt7996 drivers.

(cherry picked from commit 8a7386e787ef4d9649930bb09cdf55ccd2102346
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=b:178754244, b:337131500
TEST=Build against all mt76 drivers successfully

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5486403
Tested-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>
