+++ drivers/net/wireless/mediatek/mt76/mt7915/main.c
@@ -1727,7 +1727,7 @@ const struct ieee80211_ops mt7915_ops = {
 	.set_radar_background = mt7915_set_radar_background,
 #ifdef CONFIG_NET_MEDIATEK_SOC_WED
 	.net_fill_forward_path = mt7915_net_fill_forward_path,
-	.net_setup_tc = mt76_net_setup_tc,
+	.net_setup_tc = mt76_wed_net_setup_tc,
 #endif
 	.reconfig_complete = mt7915_reconfig_complete,
 };
